replace index with paginated posts page
